Take Into<CString> in ReaderState::new()
instead of taking &CStr and strdup'ing ourselves.

This makes the allocation clear to the user - no need to add a comment
note about it.

Also avoids an allocation in case the user already has a CString for
some reason and can pass ownership.

This is backwards-compatible, since &CStr implements this trait.