Fix Card::disconnect() keeping associated Context alive
When card.disconnect() is called manually (as opposed to being dropped),
the card is forget()ten. Since commit 4b00c65 however there is an Arc
that needs to be dropped, so do that.

In runtime this will manifest as Context.release() returning
Err(CantDispose), or silently leaking if dropped automatically.