pcsc: support direct connections to a reader
Previously, when using `ShareMode::Direct` (and `Protocols::UNDEFINED`
most likely), `connect()` will panic. This combination is useful for
connecting to a reader directly (irregardless of whether a card is
connected), and sending Control commands to it.

In this mode, `CardStatus::protocol()` panics. A new `protocol2()` is
added which returns an Option instead.

In this mode, `Card::transmit()` panics. This is likely a programmer
error, or at least a very uncommon use case. If we want to support this,
we will need a `transmit2()` with more parameters.

Fixes https://github.com/bluetech/pcsc-rust/issues/8.