Add owned versions for list_readers, get_attribute and status2
Currently, the user has to manually allocate buffers for these
functions, typically calling the associated _len function to determine
the size of the buffers.  This is handy if the user wants to manually
control allocations, but it is unnecessary complicated and error-prone
in less restricted environments.

Therefore, this patch introduces owned variants of these functions that
do not expect the buffer as an argument but instead determine the
required buffer size, allocate a buffer and return an owned version of
the data.

Fixes #17.