pcsc: handle portability hazard in Status and undeprecate it
Status was previously deprecated because it is a bitmap, which matches
pcsclite, but not Windows, where it is a simple serial enum, and cannot
be used as a bitmap.

But Status is useful, so let's instead map Windows enum to a bitmap
instead, then it will work in the same way on all platforms.

This also adds support in `status2()`/`status2_owned()`.