pcsc: switch to bitflags 2
bitflags releases a major version, but we use it in the public
interface, so upgrading is a breaking change. However, the bitflags
changelog provides guidance on how to keep 2 mostly compatible with 1:
https://github.com/bitflags/bitflags/blob/main/CHANGELOG.md#200

This is not perfect, there is still some chance of breaking dependents,
but I think/hope it is OK. I think it's better than sticking with 1 only
or bumping pcsc major.

The MSRV is bumped to match bitflags.